@@ -13,9 +13,9 @@
     "less": "^4.2.0",
     "postcss": "^8.4.31",
     "sass": "^1.69.5",
-    "stylus": "^0.61.0",
-    "typescript": "^5.2.2",
-    "vite": "^4.5.0",
-    "vite-plugin-inspect": "^0.7.41"
+    "stylus": "^0.62.0",
+    "typescript": "^5.3.2",
+    "vite": "^5.0.2",
+    "vite-plugin-inspect": "^0.7.42"
   }
 }
