@@ -11,8 +11,8 @@
     "concurrently": "^8.2.2",
     "postcss": "^8.4.31",
     "sass": "^1.69.5",
-    "typescript": "^5.2.2",
-    "vite": "^4.5.0",
-    "vite-plugin-inspect": "^0.7.41"
+    "typescript": "^5.3.2",
+    "vite": "^5.0.2",
+    "vite-plugin-inspect": "^0.7.42"
   }
 }
