@@ -1,7 +1,8 @@
 {
   "name": "@spiriit/vite-plugin-svg-spritemap",
-  "version": "1.4.0",
-  "packageManager": "pnpm@8.10.2",
+  "type": "module",
+  "version": "2.0.0",
+  "packageManager": "pnpm@8.10.5",
   "description": "Generates symbol-based SVG spritemap from all .svg files in a directory",
   "author": "Spiriit",
   "license": "MIT",
@@ -28,15 +29,15 @@
   ],
   "exports": {
     ".": {
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
     },
     "./client": {
       "types": "./dist/client.d.ts"
     }
   },
   "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
+  "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "files": [
     "dist"
@@ -55,7 +56,7 @@
     "prepublishOnly": "pnpm build"
   },
   "peerDependencies": {
-    "vite": "^3.0.0 || ^4.0.0",
+    "vite": "^4.0.0 || ^5.0.0",
     "vue": "^3.0.0"
   },
   "peerDependenciesMeta": {
@@ -68,22 +69,22 @@
     "fast-glob": "^3.3.2",
     "hash-sum": "^2.0.0",
     "mini-svg-data-uri": "^1.4.4",
-    "svgo": "^3.0.2"
+    "svgo": "^3.0.4"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "1.1.0",
+    "@antfu/eslint-config": "2.0.0",
     "@types/hash-sum": "^1.0.2",
-    "@types/node": "^14.18.63",
+    "@types/node": "^20.9.4",
     "@types/xmldom": "^0.1.34",
-    "@vitejs/plugin-vue": "^4.4.1",
+    "@vitejs/plugin-vue": "^4.5.0",
     "@vitest/coverage-v8": "^0.34.6",
     "c8": "^8.0.1",
-    "eslint": "^8.53.0",
-    "playwright": "^1.39.0",
+    "eslint": "^8.54.0",
+    "playwright": "^1.40.0",
     "svg-element-attributes": "^2.1.0",
-    "tsup": "^7.2.0",
-    "typescript": "^5.2.2",
-    "vite": "^4.5.0",
+    "tsup": "^8.0.1",
+    "typescript": "^5.3.2",
+    "vite": "^5.0.2",
     "vitest": "^0.34.6",
     "vue": "^3.3.8"
   }
