@@ -1,3 +1,3 @@
-const antfu = require('@antfu/eslint-config').default
+import antfu from '@antfu/eslint-config'
 
-module.exports = antfu()
+export default await antfu()
